+++ ui/components/portfolio/PortfolioTokensTable.js
@@ -5,7 +5,7 @@ import 'ag-grid-community/styles/ag-grid.css';
 import 'ag-grid-community/styles/ag-theme-quartz.css';
 import { Paper } from '@mui/material';
 import TokenLogoAndName from '../tokens/TokenLogoAndName';
-import DefaultCell from '../tokens/table/DefaultCell';
+import DefaultCell from '../_base/table/DefaultCell';
 import { isMobile } from 'react-device-detect';
 import { GeneralContext } from '../../../contexts/general/General.Context';
 import HoldingsCell from './HoldingsCell';